Resolve dependencies from both host go.mod and module's go.mod
Previously we would only look at the host repo's go.mod or the module's
go.mod, but not both.
This changes makes it so that we always look at both available go.mod
files and add logic to prevent reading the `replace` directives in the
module's go.mod if we also have the host repo's go.mod configured.