Fix endless recurison & use ParseLax for modules
If the host go.mod file contains a `replace` directive for a module that
we are generating and the directive only changes the version, the dependency
resolution went into an infinite recursion.

Update the code so that we use strict parsing only for the host repo's
go.mod and not for the module's, as that can cause issues.