Convert `cgo_enabled` plugin config field into a boolean
`cgo_enabled` has slightly awkward semantics. The help text and the
name imply it's a boolean, but it's actually a string that requires an
integer-looking value. In some parts of the Go build defs, the value is
expected to be `1`, otherwise cgo won't be enabled; in others, any true
value is sufficient to enable cgo (although the value of `CGO_ENABLED`
that is passed through to the underlying tool isn't always guaranteed to
be `1` - it could even be whatever value is set in the plugin
configuration).

Explicitly define `cgo_enabled` as a boolean whose default value is
false, matching the current semantics, and ensure that the `CGO_ENABLED`
environment variable is set to `1` whenever `cgo_enabled` is true.