please_go: derive Go toolchain header location from Go tool path (#276)
`go/build.Default.GOROOT` is the `GOROOT` of the Go toolchain that built
please_go. We can't rely on it existing when please_go itself runs
outside of the environment in which it was built. Instead, look for the
headers that would normally be provided in the Go toolchain in
`pkg/include` inside the Go toolchain we're using to build the code,
i.e. the one that the `go` tool is part of.

Fixes #137.