Add `strip_binaries` configuration option
Whether or not generated Go binaries are stripped is determined by the
build mode: `opt` strips them, `dbg` doesn't. There are cases in which
we might want to avoid stripping binaries even when building them in
`opt` mode - a notable example is to preserve the `.symtab` table, which
[govulncheck](https://pkg.go.dev/golang.org/x/vuln/cmd/govulncheck) uses
to perform a deeper analysis of whether vulnerable code is called by the
binary.

Add an option to the plugin configuration, `strip_binaries`, that causes
binaries to be stripped or not stripped regardless of the build mode. If
a value is not set for `strip_binaries`, `go_binary` behaves as it does
currently (i.e. stripping behaviour depends on the build mode).