please_go: remove Go <= 1.19 coverage generation code
Go 1.20 made breaking changes to the way in which coverage is computed.
This required the `coverageredesign` GOEXPERIMENT to be enabled when
using go-rules with Go 1.20 onwards.

Go 1.19 has now been EOL for almost a year, and the `coveragedesign`
GOEXPERIMENT will be removed from Go shortly, having been the default
for around 18 months. Remove the following features from please_go that
supported the Go <= 1.19 method of computing coverage:

* The `covervars` command, which is no longer necessary with
  `coverageredesign`.
* The generation of Go <= 1.19-compatible test main files with the
  `testmain` command.