Remove `coverageredesign` config field (#292)
Go 1.20 made breaking changes to the way in which coverage is computed.
This required the `coverageredesign` GOEXPERIMENT to be enabled when
using go-rules with Go 1.20 onwards.

Go 1.19 has now been EOL for almost a year, and the `coveragedesign`
GOEXPERIMENT will be removed from Go shortly, having been the default
for around 18 months. Remove support for the pre-`coverageredesign`
coverage method and deprecate the `coverageredesign` config field.

This increases the minimum Go version required by go-rules to 1.20.